Changed quite abit of the UI
Tweaked anchors to support mutliple resolutions.
Tweaked animations to better fit multiple resolutions.
Added a "flip" button to flip the object from one side of the plane to the other.
Adjusted sizes of UI components.
Continued editing the UI to fit a more modern look.
Troubleshooted multiple UI related bugs with unity and displaying 3d  models on a canvas.
Cleaned multiple objects throughout the editor and scripts.
Encapsulated a few related UI objects in the canus for a more clear understanding of what is controlling what.